// packages/language/src/generated/ast.ts
export interface AxiomBlock extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'AxiomBlock';
    name?: SectionName;
    sentences: Array<SectionSentence>;
}

export const AxiomBlock = {
    $type: 'AxiomBlock',
    name: 'name',
    sentences: 'sentences'
} as const;
// ...
export interface DefinitionBlock extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'DefinitionBlock';
    name?: SectionName;
    sentences: Array<SectionSentence>;
}

export const DefinitionBlock = {
    $type: 'DefinitionBlock',
    name: 'name',
    sentences: 'sentences'
} as const;

export function isDefinitionBlock(item: unknown): item is DefinitionBlock {
    return reflection.isInstance(item, DefinitionBlock.$type);
}

export interface Directive extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Directive';
    content: Array<DirectiveToken>;
}

export const Directive = {
    $type: 'Directive',
    content: 'content'
} as const;

export function isDirective(item: unknown): item is Directive {
    return reflection.isInstance(item, Directive.$type);
}

export interface DirectiveToken extends langium.AstNode {
    readonly $container: Directive;
    readonly $type: 'DirectiveToken';
    value: ',' | '.' | '|' | number | string;
}

export const DirectiveToken = {
    $type: 'DirectiveToken',
    value: 'value'
} as const;

export function isDirectiveToken(item: unknown): item is DirectiveToken {
    return reflection.isInstance(item, DirectiveToken.$type);
}

export interface LetBinding extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'LetBinding';
    body: Array<BodyToken>;
}

export const LetBinding = {
    $type: 'LetBinding',
    body: 'body'
} as const;

export function isLetBinding(item: unknown): item is LetBinding {
    return reflection.isInstance(item, LetBinding.$type);
}

export interface Model extends langium.AstNode {
    readonly $type: 'Model';
    elements: Array<TopLevel>;
}

export const Model = {
    $type: 'Model',
    elements: 'elements'
} as const;

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model.$type);
}

export interface ProofBlock extends langium.AstNode {
    readonly $container: TheoremBlock;
    readonly $type: 'ProofBlock';
    method?: ProofMethod;
    sentences: Array<ProofSentence>;
}

export const ProofBlock = {
    $type: 'ProofBlock',
    method: 'method',
    sentences: 'sentences'
} as const;
// ...
export type SectionBlock = AxiomBlock | DefinitionBlock | SignatureBlock | TheoremBlock;

export const SectionBlock = {
    $type: 'SectionBlock'
} as const;

export function isSectionBlock(item: unknown): item is SectionBlock {
    return reflection.isInstance(item, SectionBlock.$type);
}

export interface SectionBodyToken extends langium.AstNode {
    readonly $container: SectionSentence;
    readonly $type: 'SectionBodyToken';
    value: '(' | ')' | ',' | 'Let' | '[' | ']' | 'by' | 'contradiction' | 'induction' | '{' | '|' | '}' | number | string;
}

export const SectionBodyToken = {
    $type: 'SectionBodyToken',
    value: 'value'
} as const;

export function isSectionBodyToken(item: unknown): item is SectionBodyToken {
    return reflection.isInstance(item, SectionBodyToken.$type);
}

export interface SectionName extends langium.AstNode {
    readonly $container: AxiomBlock | DefinitionBlock | SignatureBlock | TheoremBlock;
    readonly $type: 'SectionName';
    segments: Array<string>;
}

export const SectionName = {
    $type: 'SectionName',
    segments: 'segments'
} as const;

export function isSectionName(item: unknown): item is SectionName {
    return reflection.isInstance(item, SectionName.$type);
}

export interface SectionSentence extends langium.AstNode {
    readonly $container: AxiomBlock | DefinitionBlock | SignatureBlock | TheoremBlock;
    readonly $type: 'SectionSentence';
    body: Array<SectionBodyToken>;
}

export const SectionSentence = {
    $type: 'SectionSentence',
    body: 'body'
} as const;

export function isSectionSentence(item: unknown): item is SectionSentence {
    return reflection.isInstance(item, SectionSentence.$type);
}

export interface SignatureBlock extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'SignatureBlock';
    name?: SectionName;
    sentences: Array<SectionSentence>;
}

export const SignatureBlock = {
    $type: 'SignatureBlock',
    name: 'name',
    sentences: 'sentences'
} as const;

export function isSignatureBlock(item: unknown): item is SignatureBlock {
    return reflection.isInstance(item, SignatureBlock.$type);
}

export interface TheoremBlock extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'TheoremBlock';
    kind: TheoremKind;
    name?: SectionName;
    proof?: ProofBlock;
    sentences: Array<SectionSentence>;
}

export const TheoremBlock = {
    $type: 'TheoremBlock',
    kind: 'kind',
    name: 'name',
    proof: 'proof',
    sentences: 'sentences'
} as const;

export function isTheoremBlock(item: unknown): item is TheoremBlock {
    return reflection.isInstance(item, TheoremBlock.$type);
}

export type TheoremKind = 'Corollary' | 'Lemma' | 'Proposition' | 'Theorem';

export function isTheoremKind(item: unknown): item is TheoremKind {
    return item === 'Theorem' || item === 'Proposition' || item === 'Lemma' || item === 'Corollary';
}

export type TopLevel = Directive | LetBinding | SectionBlock;

export const TopLevel = {
    $type: 'TopLevel'
} as const;